제약조건 필터링중, 공강 요일까지 필터링 완료, camelCase -> snake_case로 변수명 수정 전

+++ docker/sql/ddl/03_create_course_table.sql
@@ -10,7 +10,7 @@ CREATE TABLE course (
     class_section VARCHAR(255) NOT NULL,
     grade INT NOT NULL,
     grade_limit VARCHAR(255),
-    online_min INT DEFAULT 0,
+    online_min DECIMAL(2,1) DEFAULT 0,
     plan_code VARCHAR(255),
     semester_id VARCHAR(255) NOT NULL,
     major_id VARCHAR(255) NOT NULL,
